Refactor ChecksumValidation as UploadInterceptor plugin
  - Convert from monkey-patching to clean plugin architecture
  - Break down monolithic upload handler into focused functions
  - Use onUploadComplete hook instead of direct status observers
  - Move ChecksumValidation.js to plugin folder structure
  - Remove direct UploaderModel prototype manipulation
  - Maintain all existing checksum generation and validation functionality
