+++ src/js/core/UploadInterceptor/index.js
@@ -6,6 +6,7 @@
  */
 
 import UploadInterceptor from './UploadInterceptor.js';
+import ChecksumValidationPlugin from './plugins/ChecksumValidation/ChecksumValidation.js';
 import UploadVolumeChecker from './plugins/UploadVolumeChecker/UploadVolumeChecker.js';
 
 // Initialize the upload interceptor system when DOM loads
@@ -13,9 +14,8 @@ window.addEventListener("load", () => {
   console.log('UploadInterceptor: Initializing upload interception system');
 
   // Register plugins
+  UploadInterceptor.register(ChecksumValidationPlugin);
   UploadInterceptor.register(UploadVolumeChecker);
-
-  // Note: ChecksumValidation plugin will be registered here after refactoring
 
   console.log('UploadInterceptor: System initialized with plugins:', UploadInterceptor.getRegisteredPlugins());
 });
