+++ src/js/index.js
@@ -5,7 +5,6 @@ import "../js/core/MetadataHierarchies.js";
 import "../js/core/OAIHarvestClient.js";
 import "../js/core/publicAccessModifier.js";
 import "../js/core/tourModifier.js";
-import "../js/core/ChecksumValidation.js";
 import "../js/core/UploadInterceptor/index.js";
 import "../js/core/PassOaiToChildren.js";
 import "../js/core/consoleModifier.js";
